Add LHC epoch display driver
This code drives a pinball score display that is bolted to the wall
in Laboratorio Hacker de Campinas (http://lhc.net.br). The display
shows the amount of seconds since the laboratory was founded and an
UTC clock.

Time is synchronized on start with NTP and eventually updated again
to keep precision. The NTP code is adapted from Arduino's examples.

There's a skeleton of an IRC bot as well, but it is currently
commented out as its command parsing (specially for ping commands)
is not working as it should. One day this bot will sit in ##lhc
(irc.freenode.net) to give information to people if the space is
open or not. Maybe even do other tricks, as threre is plenty of
space in the ROM.

The pinball display is the same model that is documented here,
in Portuguese: https://garoa.net.br/wiki/Epoch

The code is pretty messy but it's been working fine so far.