Use a timer interrupt to schedule Bg tasks
It takes some time to advance the state machine while receiving commands
over Bluetooth -- and this doesn't play well when there's a task awaiting
to be executed in the background.

Implement a very naive round robin scheduler that advances with a timer
interrupt. Values have been shamelessly copied from another project I wrote
and I'm not sure if they're optimal.